[java] refactoring of Helpers and JavaSdkPath
- `Invoke` should probably be moved to Helpers
- Let’s do `using static MonoEmbeddinator4000.Helpers` throughout
`Compilation.cs`
- Fixed some string interpolation that could be removed
- `JavaSdkPath` moved to `XamarinAndroid` class
- `JavaSdkPath` is now only calculated once, previously
`/usr/libexec/java_home` would get executed three times for `-gen=Java
-platform=Android`. This also printed the Java path times three times
to the output…
