+++ binder/Utils/XamarinAndroid.cs
@@ -127,5 +127,22 @@ namespace MonoEmbeddinator4000
         {
             get { return platformDirectory.Value; }
         }
+
+        static Lazy<string> javaSdkPath = new Lazy<string>(() =>
+        {
+            // If we are running on macOS, invoke java_home to figure out Java path.
+            if (Platform.IsMacOS)
+                return Helpers.Invoke("/usr/libexec/java_home", null, null).StandardOutput.Trim();
+
+            string home = Environment.GetEnvironmentVariable("JAVA_HOME");
+            if (string.IsNullOrEmpty(home))
+                throw new Exception("Cannot find Java SDK: JAVA_HOME environment variable is not set.");
+            return home;
+        });
+
+        public static string JavaSdkPath
+        {
+            get { return javaSdkPath.Value; }
+        }
     }
 }
